refactor(copilot_task_submit/types): split tests to sibling tests_types.rs (#1266)
- src/copilot_task_submit/types.rs: 612 → 279 LOC
- New src/copilot_task_submit/tests_types.rs (332 LOC)

Pure refactor: 33 tests pass.
